move sdkutils to Common
