--- Tasks/Common/sdkutils/package.json
@@ -1,11 +1,0 @@
-{
-  "name": "sdkutils",
-  "version": "1.0.0",
-  "description": "Helper routine for manipulating the AWS SDK user agent and global configuration settings",
-  "main": "sdkutils.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Amazon Web Services",
-  "license": "MIT"
-}
